+++ Controls/sizeUtils.ts
@@ -12,7 +12,8 @@ import * as DOMUtil from './_utils/sizeUtils/DOMUtil';
 export {default as getDimensions} from './_utils/sizeUtils/getDimensions';
 export {getTextWidth} from './_utils/sizeUtils/getTextWidth';
 export {getWidth} from './_utils/sizeUtils/getWidth';
-
+export {default as ResizeObserverUtil} from './_utils/sizeUtils/ResizeObserverUtil';
+export {default as IntersectionObserver} from './_utils/sizeUtils/IntersectionObserver';
 export {
     DOMUtil
 };
